WIP: Split SW Query and related code into separate file
