+++ plugins/modules/solarwinds_info.py
@@ -242,22 +242,11 @@ from ansible.module_utils._text import to_native
 from ansible_collections.anophelesgreyhoe.solarwinds.plugins.module_utils.solarwinds_client import (
     SolarWindsClient,
     solarwindsclient_argument_spec,
-    SolarWindsQuery,
 )
 
-# TODO: Delete
-# from ansible_collections.anophelesgreyhoe.solarwinds.plugins.module_utils.model import (
-#     Model,
-# )
-
-# from ansible_collections.anophelesgreyhoe.solarwinds.plugins.module_utils.query_builder import (
-#     QueryBuilder,
-# )
-
-# TODO: Delete
-# from ansible_collections.anophelesgreyhoe.solarwinds.plugins.module_utils.swis_query_nodes import (
-#     SwisQueryNodes,
-# )
+from ansible_collections.anophelesgreyhoe.solarwinds.plugins.module_utils.solarwinds_query import (
+    SolarWindsQuery,
+)
 
 REQUESTS_IMPORT_ERROR = None
 try:
@@ -340,7 +329,6 @@ class SolarWindsInfo(object):
             query_res = self.solarwinds.client.query(query)
         except Exception as ex:
             module.fail_json(msg="Node agent query failed: {0}".format(str(ex)))
-        # module.fail_json(msg=str(query_res))
         if "results" in query_res and query_res["results"]:
             return query_res["results"]
         else:
@@ -377,240 +365,6 @@ class SolarWindsInfo(object):
         query_res = query.execute()
         return query_res
 
-        # filters = {}
-        # if "filters" in module.params and module.params["filters"] != {}:
-        #     filters = module.params["filters"]
-        #     extra_tables_filtering = [
-        #         t
-        #         for t in module.params["filters"].keys()
-        #         if t.lower() != base_table.lower()
-        #     ]
-
-        # # module.fail_json(msg="{0}".format(str(filters)))
-
-        # if "columns" in module.params and module.params["columns"] != []:
-        #     extra_tables_projecting = [
-        #         t
-        #         for t in module.params["columns"].keys()
-        #         if t.lower() != base_table.lower()
-        #     ]
-
-        # extra_tables = list(set(extra_tables_filtering) | set(extra_tables_projecting))
-        # query_tables = [base_table] + extra_tables
-
-        # module.fail_json(msg="{0}".format(str(query_tables)))
-
-        # properties = self.solarwinds.properties(module, query_tables)
-        # module.fail_json(msg="Properties: {0}".format(str(properties)))
-
-        # relationships = self.solarwinds.relationships(
-        #     module,
-        #     base_table,
-        #     [t for t in properties if t.lower() != base_table.lower()],
-        # )
-
-        # module.fail_json(msg="Properties: {0}".format(str(relationships)))
-
-        # Verify all projected columns and columns used for filtering are valid
-        projected_columns = [
-            ".".join([entity_data["aliases"][t], entity_data["properties"][t][p]])
-            for t in entity_data["all_tables"]
-            for p in list(entity_data["properties"][t].keys())
-        ]
-
-        module.fail_json(msg="Properties: {0}".format(str(projected_columns)))
-
-        projected_columns = {}
-        for table in entity_data["all_tables"]:
-            columns = []
-            for property in entity_data["properties"][table]:
-
-                # for property in list(
-                #     set(
-                #         module.params["columns"][table]
-                #         if table in module.params["columns"]
-                #         else []
-                #     )
-                #     | set(
-                #         [t for t in module.params["filters"][table]]
-                #         if table in module.params["filters"]
-                #         else []
-                #     )
-                # ):
-                if property.lower() in [p.lower() for p in properties[table]]:
-                    columns.append(
-                        ".".join(
-                            [
-                                "".join(
-                                    [u for u in properties[table] if u.isupper()]
-                                ).lower(),
-                                property,
-                            ]
-                        )
-                    )
-                else:
-                    module.fail_json(
-                        msg="Property '{0}' was not found in table '{1}'".format(
-                            property, table
-                        )
-                    )
-            projected_columns[table] = columns
-
-        # model = Model(self.solarwinds, module.params["columns"])
-        # query_columns = model.query_columns()
-
-        # module.fail_json(msg=str(projected_columns[base_table]))
-
-        query = (
-            QueryBuilder()
-            .SELECT(*projected_columns[base_table])
-            .FROM("Orion.Nodes AS n")
-        )
-
-        module.fail_json(msg="Query = {0}".format(str(query)))
-
-        if filters:
-            for filter_table in filters:
-                for filter_property in filter_table:
-                    if table != "Nodes":
-                        left, right = table_class.joins["Nodes"]
-                        query.INNER_JOIN(
-                            "{0}.{1} AS {2} ON {3}.{4} = {5}.{6}".format(
-                                table_class.schema,
-                                table,
-                                alias,
-                                alias,
-                                left,
-                                "n",
-                                right,
-                            )
-                        )
-                    table_filters = module.params["filters"][table]
-                    for column in [
-                        f for f in table_filters if f is not None and f != []
-                    ]:
-                        criteria = table_filters[column]
-                        # Filtering: translate params into a more generic criteria format to simplify querying
-                        column_criteria = ""
-                        if not isinstance(criteria, list) and not isinstance(
-                            criteria, dict
-                        ):
-                            param_value = [criteria]
-                        else:
-                            param_value = criteria
-
-                        # Iterate over each value for the current param and validate
-                        for element in [v for v in param_value if v is not None]:
-                            match = None
-                            comparator = "LIKE"
-                            wrap = "'"
-                            if (
-                                hasattr(table_class, "boolean_columns")
-                                and column in table_class.boolean_columns
-                            ):
-                                comparator = "="
-                                if isinstance(element, bool):
-                                    match = str(element)
-                                elif isinstance(element, str):
-                                    match = (
-                                        "True"
-                                        if str(element).lower() in ["yes", "on", "true"]
-                                        else "False"
-                                    )
-                                else:
-                                    module.fail_json(
-                                        msg="All filters on column '{0}' should be boolean: {1}".format(
-                                            column
-                                        )
-                                    )
-                            elif isinstance(criteria, dict):
-                                column = element
-                                match = param_value[element]
-                            elif isinstance(criteria, str):
-                                match = element
-                            elif isinstance(criteria, int):
-                                wrap = None
-                                try:
-                                    match = int(element)
-                                except Exception as ex:
-                                    pass
-                            else:
-                                match = element
-
-                            if not match:
-                                module.fail_json(
-                                    msg="Invalid filter for column '{0}'".format(column)
-                                )
-
-                            if wrap:
-                                criterion = "{0}{1}{2}".format(wrap, match, wrap)
-                            else:
-                                criterion = str(match)
-
-                            if column_criteria:
-                                column_criteria = " ".join([column_criteria, "OR"])
-                            column_criteria += " ".join(
-                                [
-                                    ".".join([alias, column]),
-                                    comparator,
-                                    criterion,
-                                ]
-                            )
-
-                            query.WHERE("({0})".format(column_criteria))
-
-        # module.fail_json(msg="{0}".format(str(query)))
-
-        # from_where = [" ".join(["FROM", " ".join(tables_spec)])]
-        # if len(criteria.strip()) > 0:
-        #     from_where.append(" ".join(["WHERE", criteria]))
-
-        # Assemble and run the query
-        # query = " ".join(["SELECT", projection] + from_where)
-
-        try:
-            query_res = self.solarwinds.client.query(str(query))
-        except Exception as ex:
-            module.fail_json(msg="Nodes query failed: {0}".format(str(ex)))
-
-        if "results" in query_res:
-            query_results = query_res["results"]
-            node_ids = [i["NodeID"] for i in query_results]
-            node_custom_properties = self.node_custom_properties(module, node_ids)
-            if node_custom_properties is not None:
-                node_custom_properties_indexed = {
-                    ncp["NodeID"]: ncp for ncp in node_custom_properties
-                }
-            else:
-                node_custom_properties_indexed = {}
-            node_agents = self.node_agents(module, node_ids)
-            if node_agents is not None:
-                node_agents_indexed = {na["NodeID"]: na for na in node_agents}
-            else:
-                node_agents_indexed = {}
-
-            nodes = []
-            for node_data in query_results:
-                if node_data["NodeID"] in node_custom_properties_indexed:
-                    node_data.update(
-                        {
-                            "CustomProperties": node_custom_properties_indexed[
-                                node_data["NodeID"]
-                            ]
-                        }
-                    )
-                if node_data["NodeID"] in node_agents_indexed:
-                    node_data.update(
-                        {"Agent": node_agents_indexed[node_data["NodeID"]]}
-                    )
-
-                nodes.append(node_data)
-
-        else:
-            nodes = None
-        info = {"nodes": nodes, "count": len(nodes), "query": str(query)}
-        return info
-
 
 # ==============================================================
 # main
@@ -625,82 +379,12 @@ def main():
         columns=dict(
             type="dict",
         ),
-        #     options=dict(
-        #         Nodes=dict(
-        #             type="list",
-        #             default=["NodeID", "Caption", "DNS", "IPAddress", "Uri"],
-        #         ),
-        #         Agents=dict(type="list", default=[]),
-        #         NodesCustomProperties=dict(type="list", default=[]),
-        #     ),
-        # ),
         include=dict(
             type="dict",
         ),
         exclude=dict(
             type="dict",
         ),
-        # options=dict(
-        #     Nodes=dict(
-        #         type="dict",
-        #         options=dict(
-        #             Caption=dict(type="list", elements="str", default=[]),
-        #             DNS=dict(type="list", elements="str", default=[]),
-        #             IPAddress=dict(type="list", elements="str", default=[]),
-        #             IsServer=dict(type="bool"),
-        #             MachineType=dict(type="list", elements="str", default=[]),
-        #             NodeDescription=dict(type="list", elements="str", default=[]),
-        #             NodeID=dict(type="list", elements="int", default=[]),
-        #             ObjectSubType=dict(
-        #                 type="list",
-        #                 elements="str",
-        #                 choices=["Agent", "ICMP", "SNMP", "WMI"],
-        #                 default=[],
-        #             ),
-        #             EngineID=dict(type="list", elements="int", default=[]),
-        #             SNMPVersion=dict(
-        #                 type="list",
-        #                 elements="int",
-        #                 choices=[0, 1, 2, 3],
-        #                 default=[],
-        #             ),
-        #         ),
-        #     ),
-        #     Agents=dict(type="dict", default={}),
-        #     NodesCustomProperties=dict(type="dict", default={})
-        # agent=dict(type="dict", default={}),
-        # caption=dict(type="list", elements="str", default=[]),
-        # custom_properties=dict(type="dict", default={}),
-        # dns=dict(type="list", elements="str", default=[]),
-        # ip_address=dict(type="list", elements="str", default=[]),
-        # is_server=dict(type="bool"),
-        # machine_type=dict(type="list", elements="str", default=[]),
-        # node_description=dict(type="list", elements="str", default=[]),
-        # node_id=dict(type="list", elements="int", default=[]),
-        # object_sub_type=dict(
-        #     type="list",
-        #     elements="str",
-        #     choices=["agent", "icmp", "snmp", "wmi"],
-        #     default=[],
-        # ),
-        # polling_engine_id=dict(type="list", elements="int", default=[]),
-        # polling_method=dict(
-        #     type="list",
-        #     elements="str",
-        #     choices=["agent", "icmp", "snmp", "wmi"],
-        #     default=[],
-        # ),
-        # severity_max=dict(type="int"),
-        # severity_min=dict(type="int"),
-        # snmp_version=dict(
-        #     type="list", elements="int", choices=[0, 1, 2, 3], default=[]
-        # ),
-        # status=dict(type="list", elements="str", default=[]),
-        # sys_name=dict(type="list", elements="str", default=[]),
-        # unmanaged=dict(type="bool"),
-        # vendor=dict(type="list", elements="str", default=[]),
-        # ),
-        # ),
     )
 
     argument_spec.update(solarwindsclient_argument_spec())
